refactor(capabilities): split capability metadata into a shared package
cli/ could not import the registry: cli/tsconfig.json sets rootDir: "src", so
pulling in server/registry dragged the whole server tree across the package
boundary — 113 TS6059 errors in the real production build, which neither tsc -b
nor vitest catches because cli/ is not in the root project graph.

handler is the only field needing server internals, so metadata now lives in
@octomux/capabilities and server composes Capability = CapabilityMeta &
{ handler }. The CLI projection moves there too and takes its capability list as
a parameter instead of importing a server-side registry.

zod is declared as a peer dependency, not a dependency. Declaring it directly
made bun re-resolve the transitive zod from @modelcontextprotocol/sdk
('^3.25 || ^4.0') from 3.25.76 up to 4.4.3, which silently broke z.infer in
orchestrator/command-registry.ts across files this change never touched.
