+++ packages/capabilities/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@octomux/capabilities",
+  "version": "0.1.0",
+  "type": "module",
+  "license": "MIT",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc -b",
+    "typecheck": "tsc -b"
+  },
+  "peerDependencies": {
+    "@octomux/api-client": "workspace:*",
+    "commander": "*",
+    "zod": "*"
+  }
+}
